Added Card.java and added card drawing functionality
Added the Card class which will allows me to store the different
blackjack values as well as the suit of the cards. Also added an
ArrayList of Cards which will hold the entire deck of 52 cards. Cards
are loaded in via the loadCards() function in BlackjackScreen. Currently
the game just loads in a random card from the deck.